refactor: Split Viewer component from index into files
no code change
